Refactor: ServiceItem, Book 도메인 책임 분리
