+++ src/main/java/teamssavice/ssavice/serviceItem/controller/ServiceItemController.java
@@ -8,6 +8,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import teamssavice.ssavice.auth.constants.Role;
+import teamssavice.ssavice.book.controller.dto.BookResponse;
+import teamssavice.ssavice.book.service.dto.BookModel;
 import teamssavice.ssavice.global.annotation.CurrentId;
 import teamssavice.ssavice.global.annotation.RequireRole;
 import teamssavice.ssavice.global.dto.CursorResult;
@@ -85,15 +87,15 @@ public class ServiceItemController {
 
     @PostMapping("/{serviceId}/apply")
     @RequireRole(Role.USER)
-    public ResponseEntity<ServiceItemResponse.Apply> applyServiceItem(
+    public ResponseEntity<BookResponse.Apply> applyServiceItem(
             @CurrentId Long userId,
             @PathVariable Long serviceId
     ) {
 
         ServiceItemCommand.Apply command = ServiceItemCommand.Apply.of(userId, serviceId);
 
-        ServiceItemModel.Apply model = serviceItemService.apply(command);
+        BookModel.Apply model = serviceItemService.apply(command);
 
-        return ResponseEntity.ok(ServiceItemResponse.Apply.from(model));
+        return ResponseEntity.ok(BookResponse.Apply.from(model));
     }
 }
